refactor: moved update workout form to own component
